+++ packages/types/src/api/web/workspace/automation.ts
